Upgrade to version 1.1.0 of https://github.com/khatchad/Common-Eclipse-Refactoring-Framework.
