+++ edu.cuny.citytech.foreachlooptolambda.ui/src/edu/cuny/citytech/foreachlooptolambda/ui/contributions/ForeachLoopToLambdaRefactoringContribution.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import org.eclipse.ltk.core.refactoring.RefactoringDescriptor;
 
 import edu.cuny.citytech.foreachlooptolambda.ui.descriptors.ForeachLoopToLambdaRefactoringDescriptor;
-import edu.cuny.citytech.refactoring.common.RefactoringContribution;
+import edu.cuny.citytech.refactoring.common.core.RefactoringContribution;
 
 public class ForeachLoopToLambdaRefactoringContribution
 		extends RefactoringContribution {
