+++ edu.cuny.citytech.foreachlooptolambda.ui/src/edu/cuny/citytech/foreachlooptolambda/ui/descriptors/ForeachLoopToLambdaRefactoringDescriptor.java
@@ -6,8 +6,8 @@ package edu.cuny.citytech.foreachlooptolambda.ui.descriptors;
 import java.util.Map;
 
 import edu.cuny.citytech.foreachlooptolambda.ui.refactorings.ForeachLoopToLambdaRefactoring;
-import edu.cuny.citytech.refactoring.common.Refactoring;
-import edu.cuny.citytech.refactoring.common.RefactoringDescriptor;
+import edu.cuny.citytech.refactoring.common.core.Refactoring;
+import edu.cuny.citytech.refactoring.common.core.RefactoringDescriptor;
 
 /**
  * @author raffi
