+++ edu.cuny.citytech.foreachlooptolambda.ui/src/edu/cuny/citytech/foreachlooptolambda/ui/refactorings/ForeachLoopToLambdaRefactoring.java
@@ -19,7 +19,7 @@ import org.eclipse.ltk.core.refactoring.RefactoringStatus;
 
 import edu.cuny.citytech.foreachlooptolambda.ui.messages.Messages;
 import edu.cuny.citytech.foreachlooptolambda.ui.visitor.LambdaConversionVisitor;
-import edu.cuny.citytech.refactoring.common.Refactoring;
+import edu.cuny.citytech.refactoring.common.core.Refactoring;
 
 /**
  * The activator class controls the plug-in life cycle
